Merge branch 'bnxt_en-NS2-Nitro'
Michael Chan says:

====================
bnxt_en: Add support for NS2 Nitro.

This series adds support for the embedded version of the
ethernet controller (Nitro) in the North Star 2 SoC.  There are a number
of features not supported and a software workaround for a hardware rx
bug is required for Nitro A0.  Please review.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: dfa15378bd69c6b193a483725889dd90c1a7dd48 ]