Merge branch 'bridge-vlan-stats-with-flags'
Nikolay Aleksandrov says:

====================
net: bridge: export vlan stats per-port with flags

This set adds the ability to export vlan stats per-port. Patch 01 makes
that possible by consolidating the bridge and port linkxstats calls. Then
patch 02 allows to dump the vlan entry flags in order to be able to
distinguish between bridge and port vlan entries when dumping the master
device vlan stats. That is needed because that call was implemented when
the stats API didn't have slave dumping capabilities and it dumps all vlan
stats (for both bridge and port entries). We also need it in order to print
the vlan flags when dumping the stats.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 53e43f3905f44366acbb5ff6f295d64996fbf6a3 ]