ARM: dts: sun5i: Add dts file for the Empire Electronix M712 tablet
Add a dts file for the Empire Electronix M712 tablet, this is a 7" A13
tablet, with micro-usb (otg), headphone and micro-sd slots on the
outside. It uses a Goodix gt811 touchscreen controller, a RTL8188CTV
wifi chip and a DMART06 (1238a4) accelerometer.

Note currently the dts for this is somewhat empty. This will change
once we add support for the touchscreen and accelerometer.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>

[ upstream commit: cb0db11afe72eb5206ba35cd9afa2868d0550cbe ]