Merge branch 'mlxsw-fw-mark-offload'
Jiri Pirko says:

====================
mlxsw: Introduce support for offload forward mark

Ido says:
This patchset enables the forwarding of certain control packets by the
device instead of relying on the CPU to do the forwarding.

The first two patches simplify the current switchdev offload forward
infrastructure and make it usable for stacked devices. This is done by
moving the packet and port marking to the bridge driver instead of the
switch driver.

Patches 3-5 add the mlxsw specific bits to support the forward mark.
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: ed35ca994f97211abdbc87063d8869ead99b508c ]