ARM: dts: add device tree for the RealView EB Rev D
The RealView EB baseboard revision D is a special case as it appears
to be undocumented in ARM official documents, while the Linux kernel
still contains special code for handling it.

commit be4f3c8691492934c8ee03dbecb3a3a865ac6cd6
"Add RealView/EB support for the LAN9118 Ethernet chip"
adds support for the SMSC LAN9118 chip used on the D revision of
the baseboard, but no other traces of hardware deviations for this
variant can be found.

This creates a separate top-level .dts file especially for this
board version, so that it gets registered with the right type
of ethernet controller. The ethernet controller modifications
are then put into a separate .dtsi file so it can be overlaid
on other EB variants using the revision D baseboard.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

[ upstream commit: 6a29fa31cdcf9e2ddf74406bd084067b4917edcd ]